add: health ui

+++ Assets/Scripts/Components/MeleeAttackComponent.cs
@@ -8,7 +8,7 @@ namespace Components
     [Serializable][Documentation(Doc.NONE, "")]
     public sealed class MeleeAttackComponent : BaseComponent
     {
-        [field: SerializeField] public float attackPower { get; private set; } = 1;
+        [field: SerializeField] public int attackPower { get; private set; } = 1;
         [field: SerializeField] public float attackRangeRadius { get; private set; } = 1;
         [field: SerializeField] public LayerMask enemyLayerMask { get; private set; }
     }
